refactor(users): standardize users endpoints and files naming
