+++ library_service_api/urls.py
@@ -2,7 +2,7 @@ from django.urls import path, include
 from rest_framework import routers
 
 from library_service_api.views import BookViewSet
-from library_service_auth.views import CreateCustomerView
+from library_service_users.views import CreateCustomerView
 
 app_name = "library_service_api"
