+++ library_service_users/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class LibraryServiceAuthConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'library_service_auth'
+    name = 'library_service_users'
